pyproject.toml
[project]
name = "precognito"
version = "0.1.0"
description = "IoT-enabled Predictive Maintenance System"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # Data Processing
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "influxdb-client>=1.40.0",
    # IoT/MQTT
    "paho-mqtt>=2.1.0",
    # ML/AI
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    # Utils
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "slowapi>=0.1.9",
    "pytest>=9.0.2",
    "psycopg2-binary>=2.9.11",
    "requests>=2.33.1",
    "alembic>=1.18.4",
    "python-json-logger>=4.1.0",
]

[tool.uv]
exclude-dependencies = [
    "nvidia-nccl-cu12", 
    "nvidia-cuda-runtime-cu12", 
    "nvidia-cuda-cupti-cu12", 
    "nvidia-cudnn-cu12", 
    "nvidia-cublas-cu12",
    "nvidia-cuda-nvrtc-cu12",
    "nvidia-nvjitlink-cu12",
    "nvidia-curand-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-cusparse-cu12",
    "nvidia-nvtx-cu12"
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-env>=1.6.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest-mock>=3.15.1",
    "pytest-env>=1.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["tests"]
asyncio_mode = "auto"
env = [
    "DATABASE_URL=sqlite:///./test_precognito.sqlite",
    "INFLUX_URL=http://localhost:8086",
    "INFLUX_TOKEN=test_token",
    "INFLUX_ORG=test_org",
    "INFLUX_BUCKET=test_bucket",
    "MQTT_BROKER=localhost",
    "BETTER_AUTH_SECRET=test_secret"
]
